Lay the foundation of the application
This commit is the first major step forward. This commit adds the entire grid-layout system, a modular way of rendering and handling settings for widgets, as well as a clock widget. Each widget is saved to localstorage along with it's settings and current state. A widget works like a module; a widget is self-contained within a file in the widgets folder (apart from the hard-coded work in the widgets.js file... still trying to find a solution for that), which defines how the widget is rendered, and it's default settings. Any widget can be created from a button in the main component (still trying to find a truely modular solution, as currently those buttons are hard-coded as well).
The next step is to work on some more widgets to make sure everything works as it should.