Add Calendar widget, rebuild Google API code
Ouch calendars hurt. Moved the initializing code for the Google API from the Gmail widget to the gridController so that any widget can access it without creating it's own seperate client. I'm taking a step away from the calendar for now, but there are still some improvements I would like to make, such as inspecting an item in detail and maybe selecting which calendars to display. Right now the calendar just pulls whatever the user has selected as visible on the website and grabs all the events associated with those calendars. Speaking of which, this commit also adds a little todo list to the readme, instead of keeping the list locally.