Begin working on write functionality
THIS COMMIT DOES NOT WORK! DO NOT TRY TO RUN THIS VERSION.
This marks a shift in the scope of this project. This project will now focus on mainly read-only information, and only accept very limited input from the user. This commit was an experiment in investigating how taking input could be best done, but after working on it for awhile, it only brought up more inconveniences for little benefit. For example, because the user would be able to click and drag within the widgets, they would only be able to drag the widget from the edges. Also, as of current plans, only a tiny portion of the widgets would use direct writing functionality. The alternative is what is currently in use for the notepad widget, putting the control in the settings. It's a trade off that uses the lesser of two evils.
This commit was made to save the working code in case this functionality should ever be implemented in the future. This is the beginnings of working code, the work just needs to be completed.