Begin working on recurrence rules for calendar
Also putting this project on indefinite hold. There are too many problems coming up that are making this project not worth continuing. It also doesn't help that as Google is rolling out their major updates to their web services like calendar and Tasks, their public interfaces are being left behind. Their new companion that they've integrated into the sides of GMail, Drive, and Calendar are good enough for now for my uses, but unfortunately whatever magic they've used to get it to work can't be reused elsewhere.

It's unfortunate, but because of this and the problems surrounding rrules make it easier to use whats already available rather than to make this from scratch. Maybe once google updates their stuff it will be easier to work with in a project like this, but until then I don't think it's worth doing.